Bugfix: Only create .configfiles.bak if it doesn’t already exist

+++ initialize.sh
@@ -36,7 +36,10 @@ link_dir() {
 
 		# Preserve the file, if it already exists
 		[[ -e "$DEST" && ! -L "$DEST" ]] && {
-			mkdir -p "$HOME/.configfiles.bak" || exit 1
+			if [ ! -f "$HOME/.configfiles.bak" ]
+			then
+				mkdir -p "$HOME/.configfiles.bak" || exit 1
+			fi
 			mv "$DEST" "$HOME/.configfiles.bak/"
 		}
